fix(ui): fix timestamp formatting (#551)
When filtering topic data by `timestamp` and also when fetching offsets
for consumer groups by `datetime` we seem to be formatting the hour of
the timestamp using `hh`. In moment.js, this causes the hour to always
have a range between [0-12](https://momentjs.com/docs/#/displaying/format/)
even though even though the api expects a range between 0-24. This prevents
filtering of timestamp or fetching of offsets for any time greater than 12:59
on any given day.

This PR corrects this by changing `hh` to `HH` on the formatting calls.

close #550